requirements.txt
streamlit>=1.50.0
deepface
opencv-python-headless
mediapipe
tensorflow-cpu
tf-keras
Pillow